requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2.3
django-extensions==3.1.3
django-nested-inlines==0.1
django-simpleui==2021.6.1
Faker==8.4.0
faker-vehicle==0.2.0
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
